style: align wrapped index expression and section spacing
Agent-Logs-Url: https://github.com/lauradouglas/flixie-app/sessions/bce3d2e2-e7ca-4524-a58c-4ab705039293

+++ lib/screens/wrapped_screen.dart
@@ -130,7 +130,7 @@ class _WrappedViewState extends State<_WrappedView> {
                     ...wrapped.monthlyWatchCounts.map(
                       (m) => ListTile(
                         dense: true,
-                        title: Text(_monthNames[(m.month - 1).clamp(0, 11).toInt()]),
+                        title: Text(_monthNames[(m.month - 1).clamp(0, 11)]),
                         trailing: Text('${m.count}'),
                       ),
                     ),
